requirements.txt
pyperclip~=1.9.0
pillow~=11.0.0